+++ CHANGELOG.md
@@ -16,6 +16,7 @@
 - `Innmind\Cli\Question\ChoiceQuestion` now return the read data in an `Innmind\Immutable\Attempt`
 - `Innmind\Cli\Question\Question` now return an `Innmind\Immutable\Attempt`
 - `Innmind\Cli\Question\ChoiceQuestion` now return an `Innmind\Immutable\Attempt`
+- `Innmind\Cli\Commands` now return an `Innmind\Immutable\Attempt`
 
 ### Fixed